Renderer2D reached a usable state
For now, Renderer2D displays 1x1 Spritesheets inefficiently. Further development will seek to optimize the rendering of Drawable structures, utilising common SpriteSheets(and possibly queues). Another priority is getting the SpriteSheets to work as intended.

Note: A decision needs to be made regarding the placement of the main loop. If it is outside, then that requires that the GLFWwindow handle be publically accessible.

Changes:
-Basic functionality added and tested.
-Set up proper initialisation routine for Renderer2D.
-StandardQuad structure moved to main_renderer.h since it'll only be used by it.
-Added methods for setting common uniforms(mat4 and floats)
-Small fix in Camera::ScreenToWorldSpaceConversion()

+Added testing shader
+Needs any .PNG image file for tests