InputController implemented
Changes:
-Component InputController added, implemented and tested
--Uses 2 std::bitsets to track a range of keystates defined at initialisation
--Supports easy checking if a button has just been pressed, is held, or just released.

Note: This works with GLFW_KEY_ codes. They are not ideal and will be replaced in the future, possibly with an enumerated class.
Note: The InputController tracking and capturing of keystates has ranges of indexes which are never used. Need to look over possibility of caching without heap allocations.